refactor: satisfy UI control CI gates
